requirements.txt
# Neural Trading Pipeline - TDD Requirements
# Python 3.10+

# Core ML Framework
tensorflow>=2.15.0
keras>=3.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# CCXT for exchange data
ccxt>=4.0.0

# Configuration Management
pyyaml>=6.0.1
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Technical Indicators
ta-lib>=0.4.0

# Backtesting (future)
# backtrader>=1.9.78.123

# Model Persistence
joblib>=1.3.0
h5py>=3.10.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
pylint>=2.17.5

# Visualization
matplotlib>=3.7.2
seaborn>=0.12.2

# Experiment Tracking
tensorboard>=2.15.0
# mlflow>=2.7.0  # Optional
# wandb>=0.15.0  # Optional

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.2

# Performance
psutil>=5.9.5

# Development
ipython>=8.14.0
jupyter>=1.0.0
jupyterlab>=4.0.0

# Monitoring (future)
# prometheus-client>=0.17.1

# API (future)
# fastapi>=0.103.0
# uvicorn>=0.23.0
# pydantic>=2.3.0
